r_address refactor
