+++ xrpl-wasm-stdlib/xrpl-macros/Cargo.toml
@@ -23,3 +23,6 @@ features = ["full"]
 
 [dependencies.quote]
 version = "1.0"
+
+[dev-dependencies]
+trybuild = "1.0"
