+++ xrpl-wasm-stdlib/xrpl-macros/tests/r_address/fail_non_literal.rs
@@ -1,0 +1,10 @@
+use xrpl_macros::r_address;
+
+// The macro only accepts string literals, never runtime expressions.
+fn convert(addr: &str) {
+    r_address!(addr);
+}
+
+fn main() {
+    convert("rHb9CJAWyB4rj91VRWn96DkukG4bwdtyTh");
+}
