+++ xrpl-wasm-stdlib/xrpl-macros/tests/r_address/fail_non_literal.stderr
@@ -1,0 +1,5 @@
+error: expected string literal
+ --> tests/r_address/fail_non_literal.rs:5:16
+  |
+5 |     r_address!(addr);
+  |                ^^^^
